# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Portable Codex Relay
  version: 0.1.0
  description: >-
    REST broker for a ChatGPT Custom GPT to request file operations on a paired desktop companion.
    The relay forwards JSON to the device and never reads or writes disk. Workspace trust, path sandboxing,
    credential checks, and optional write-approval prompts are enforced entirely on the companion; treat
    every Action call as non-consequential at the Chat layer because the device is the authority that allows or denies work.
servers:
  - url: https://YOUR-RELAY-URL.example.com
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  schemas:
    ToolError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
    ApprovalRequired:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
      additionalProperties: false
    ListTrustedWorkspacesRequest:
      type: object
      properties:
        requestId:
          type: string
          description: Optional caller-generated request ID. The relay generates one if omitted.
        deviceId:
          type: string
          description: Optional target device ID. Defaults to the API token's default device.
      additionalProperties: true
    GetGptInstructionsRequest:
      type: object
      properties:
        requestId:
          type: string
          description: Optional caller-generated request ID. The relay generates one if omitted.
        deviceId:
          type: string
          description: Optional target device ID. Defaults to the API token's default device.
      additionalProperties: true
    ListSkillsRequest:
      type: object
      properties:
        requestId:
          type: string
          description: Optional caller-generated request ID. The relay generates one if omitted.
        deviceId:
          type: string
          description: Optional target device ID. Defaults to the API token's default device.
      additionalProperties: true
    GetSkillRequest:
      type: object
      properties:
        requestId:
          type: string
          description: Optional caller-generated request ID. The relay generates one if omitted.
        deviceId:
          type: string
          description: Optional target device ID. Defaults to the API token's default device.
        skillName:
          type: string
          description: Skill name to load. Use the text after / when the user activates /skill-name.
        maxBytes:
          type: integer
          description: Maximum UTF-8 bytes of SKILL.md instructions to return.
          minimum: 1
      required:
        - skillName
      additionalProperties: true
...
    WriteStdinRequest:
      type: object
      properties:
        requestId:
          type: string
        deviceId:
          type: string
        processId:
          type: string
          description: Running process identifier from a Codex-style shell session.
        sessionId:
          type: string
          description: Alias for processId used by some harnesses.
        session_id:
          type: integer
          description: Codex-style numeric session identifier returned by exec_command.
        chars:
          type: string
          description: Characters to write to stdin. May be empty to poll for output.
        yield_time_ms:
          type: integer
          description: How long to wait for output before returning.
          minimum: 0
        max_output_tokens:
          type: integer
          description: Approximate Codex-style output budget. The companion converts this to a byte limit.
          minimum: 1
        stdin:
          type: string
        input:
          type: string
          description: Alias for stdin.
      additionalProperties: true
    RequestPermissionsRequest:
      type: object
      properties:
        requestId:
          type: string
        deviceId:
          type: string
        permissions:
          type: array
          items:
            type: string
        reason:
          type: string
      additionalProperties: true
...
    ToolResponse:
      type: object
      description: >-
        Relay response from the paired device. Trust and filesystem access are decided on the device;
        `status` may be `denied` or include `approvalRequired` when the companion blocks or defers a change.
      required:
        - requestId
        - status
      properties:
        requestId:
          type: string
        status:
          type: string
          enum:
            - ok
            - denied
            - error
            - timeout
        result:
          description: Tool-specific payload when status is ok.
        error:
          $ref: "#/components/schemas/ToolError"
        approvalRequired:
          $ref: "#/components/schemas/ApprovalRequired"
      additionalProperties: true
paths:
  /tools/list-trusted-workspaces:
    post:
      x-openai-isConsequential: false
      operationId: list_trusted_workspaces
      summary: List the trusted workspace roots available on the connected device.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListTrustedWorkspacesRequest"
      responses:
        "200":
          description: Tool result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
  /tools/get-gpt-instructions:
    post:
      x-openai-isConsequential: false
      operationId: get_gpt_instructions
      summary: Return the latest Custom GPT instructions from the paired companion.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetGptInstructionsRequest"
      responses:
        "200":
          description: Tool result with result.instructions containing the current instruction text.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
  /tools/list-skills:
    post:
      x-openai-isConsequential: false
      operationId: list_skills
      summary: List imported Codex CLI style skills available on the connected device.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListSkillsRequest"
      responses:
        "200":
          description: Tool result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
  /tools/get-skill:
    post:
      x-openai-isConsequential: false
      operationId: get_skill
      summary: Load a Codex CLI style skill by name and return its SKILL.md instructions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetSkillRequest"
      responses:
        "200":
          description: Tool result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
...
  /tools/write-stdin:
    post:
      x-openai-isConsequential: false
      operationId: write_stdin
      summary: Write characters to an existing exec_command session and return recent output.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteStdinRequest"
      responses:
        "200":
          description: Tool result with Codex-style output fields such as session_id, exit_code, wall_time_seconds, and output.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
  /tools/request-permissions:
    post:
      x-openai-isConsequential: false
      operationId: request_permissions
      summary: Codex compatibility shim for sandbox permission requests.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPermissionsRequest"
      responses:
        "200":
          description: Tool result. Portable Codex uses trusted workspaces and companion approvals instead.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
...